requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==1.53.0
PyPDF2==3.0.1
pandas==2.2.0
openpyxl==3.1.2
pillow==10.4.0
pytesseract==0.3.13
python-docx==1.1.2
numpy==1.26.0
scikit-learn==1.4.0
python-dotenv==1.0.0
werkzeug==2.3.7